Implement _get_extra_facets as method of Dataset
