+++ tests/unit/config/test_config.py
@@ -8,9 +8,94 @@ import yaml
 from esmvalcore.cmor.check import CheckLevels
 from esmvalcore.config import CFG, _config, _config_validators
 from esmvalcore.config._config import (
+    _deep_update,
     get_ignored_warnings,
+    load_extra_facets,
 )
 from esmvalcore.exceptions import ESMValCoreDeprecationWarning, RecipeError
+
+TEST_DEEP_UPDATE = [
+    ([{}], {}),
+    ([{"a": 1, "b": 2}, {"a": 3}], {"a": 3, "b": 2}),
+    (
+        [
+            {"a": {"b": 1, "c": {"d": 2}}, "e": {"f": 4, "g": 5}},
+            {"a": {"b": 2, "c": 3}},
+        ],
+        {"a": {"b": 2, "c": 3}, "e": {"f": 4, "g": 5}},
+    ),
+]
+
+
+# TODO: remove in v2.15.0
+@pytest.mark.parametrize(("dictionaries", "expected_merged"), TEST_DEEP_UPDATE)
+def test_deep_update(dictionaries, expected_merged):
+    merged = dictionaries[0]
+    for update in dictionaries[1:]:
+        merged = _deep_update(merged, update)
+    assert expected_merged == merged
+
+
+BASE_PATH = importlib_files("tests") / "sample_data" / "extra_facets"
+
+TEST_LOAD_EXTRA_FACETS = [
+    ("test-nonexistent", (), {}),
+    ("test-nonexistent", (BASE_PATH / "simple",), {}),  # type: ignore
+    (
+        "test6",
+        (BASE_PATH / "simple",),  # type: ignore
+        {
+            "PROJECT1": {
+                "Amon": {
+                    "tas": {
+                        "cds_var_name": "2m_temperature",
+                        "source_var_name": "2t",
+                    },
+                    "psl": {
+                        "cds_var_name": "mean_sea_level_pressure",
+                        "source_var_name": "msl",
+                    },
+                },
+            },
+        },
+    ),
+    (
+        "test6",
+        (BASE_PATH / "simple", BASE_PATH / "override"),  # type: ignore
+        {
+            "PROJECT1": {
+                "Amon": {
+                    "tas": {
+                        "cds_var_name": "temperature_2m",
+                        "source_var_name": "t2m",
+                    },
+                    "psl": {
+                        "cds_var_name": "mean_sea_level_pressure",
+                        "source_var_name": "msl",
+                    },
+                    "uas": {
+                        "cds_var_name": "10m_u-component_of_neutral_wind",
+                        "source_var_name": "u10n",
+                    },
+                    "vas": {
+                        "cds_var_name": "v-component_of_neutral_wind_at_10m",
+                        "source_var_name": "10v",
+                    },
+                },
+            },
+        },
+    ),
+]
+
+
+# TODO: Remove in v2.15.0
+@pytest.mark.parametrize(
+    ("project", "extra_facets_dir", "expected"),
+    TEST_LOAD_EXTRA_FACETS,
+)
+def test_load_extra_facets(project, extra_facets_dir, expected):
+    extra_facets = load_extra_facets(project, extra_facets_dir)
+    assert extra_facets == expected
 
 
 def test_get_project_config(mocker):
